Replace UserCluster Prometheus with Grafana Agent
