+++ pkg/controller/user-cluster-controller-manager/resources/controller.go
@@ -65,10 +65,10 @@ const (
 )
 
 type UserClusterMLA struct {
-	Logging                      bool
-	Monitoring                   bool
-	MLAGatewayURL                string
-	PrometheusScrapeConfigPrefix string
+	Logging                           bool
+	Monitoring                        bool
+	MLAGatewayURL                     string
+	MonitoringAgentScrapeConfigPrefix string
 }
 
 // Add creates a new user cluster controller.
